logl and expl modif for overflow
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/exactsampling@3 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/HyperGeometrique.cpp
@@ -19,10 +19,10 @@ HyperGeometrique::HyperGeometrique(int N)
 {
 	Log_n_factoriel = new long double[N];
 	Log_n_factoriel[0] = 0; // log(1)
-	Log_n_factoriel[1] = logl(2);  // log(2)
+	Log_n_factoriel[1] = logl(2.);  // log(2)
 	for (int i = 2; i < N ; i++)
 	{
-		Log_n_factoriel[i] = Log_n_factoriel[i - 1] + logl(i + 1);
+		Log_n_factoriel[i] = Log_n_factoriel[i - 1] + logl(static_cast<double>(i + 1));
 	}
 	
 }
